requirements.txt
asn1crypto==0.23.0
astroid==1.5.3
autopep8==1.3.2
Cerberus==1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.3
Django==1.11.4
django-cors-headers==2.1.0
django-enum-field==0.2
django-extensions==1.8.1
django-filter==1.0.4
django-mysql==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
h2==2.6.2
hpack==3.0.0
httplib2==0.10.3
hyper==0.7.0
hyperframe==3.2.0
idna==2.5
isort==4.2.15
itypes==1.1.0
lazy-object-proxy==1.3.1
Markdown==2.6.8
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
mysqlclient==1.3.10
olefile==0.44
openapi-codec==1.3.2
pbr==3.1.1
psycopg2==2.7.3
py==1.4.34
pycodestyle==2.3.1
pycparser==2.18
PyJWT==1.5.3
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
PySocks==1.6.7
pytest==3.2.1
pytest-cov==2.5.1
pytest-runner==2.12
python-enumeration==0.1
pytz==2017.2
requests==2.18.3
simplejson==3.11.1
urllib3==1.22
webencodings==0.5.1
wrapt==1.10.11
